+++ java/com/thebombzen/jxlatte/color/OpsinInverseMatrix.java
@@ -9,27 +9,27 @@ import com.thebombzen.jxlatte.util.MathHelper;
 
 public class OpsinInverseMatrix {
 
-    private static final double[][] DEFAULT_MATRIX = {
-        {11.031566901960783D, -9.866943921568629D, -0.16462299647058826D},
-        {-3.254147380392157D, 4.418770392156863D, -0.16462299647058826D},
-        {-3.6588512862745097D, 2.7129230470588235D, 1.9459282392156863D}
+    private static final float[][] DEFAULT_MATRIX = {
+        {11.031566901960783f, -9.866943921568629f, -0.16462299647058826f},
+        {-3.254147380392157f, 4.418770392156863f, -0.16462299647058826f},
+        {-3.6588512862745097f, 2.7129230470588235f, 1.9459282392156863f}
     };
 
-    private static final double[] DEFAULT_OPSIN_BIAS = {
-        -0.0037930732552754493D, -0.0037930732552754493D, -0.0037930732552754493D
+    private static final float[] DEFAULT_OPSIN_BIAS = {
+        -0.0037930732552754493f, -0.0037930732552754493f, -0.0037930732552754493f
     };
 
-    private static final double[] DEFAULT_QUANT_BIAS = {
-        1D - 0.05465007330715401D, 1D - 0.07005449891748593D, 1D - 0.049935103337343655D
+    private static final float[] DEFAULT_QUANT_BIAS = {
+        1f - 0.05465007330715401f, 1f - 0.07005449891748593f, 1f - 0.049935103337343655f
     };
 
-    private static final double DEFAULT_QBIAS_NUMERATOR = 0.145D;
+    private static final float DEFAULT_QBIAS_NUMERATOR = 0.145f;
 
-    private double[][] matrix;
-    private double[] opsinBias;
-    private double[] cbrtOpsinBias;
-    public final double[] quantBias;
-    public final double quantBiasNumerator;
+    private float[][] matrix;
+    private float[] opsinBias;
+    private float[] cbrtOpsinBias;
+    public final float[] quantBias;
+    public final float quantBiasNumerator;
     public final CIEPrimaries primaries;
     public final CIEXY whitePoint;
 
@@ -54,16 +54,16 @@ public class OpsinInverseMatrix {
             quantBias = DEFAULT_QUANT_BIAS;
             quantBiasNumerator = DEFAULT_QBIAS_NUMERATOR;
         } else {
-            matrix = new double[3][3];
+            matrix = new float[3][3];
             for (int i = 0; i < 3; i++) {
                 for (int j = 0; j < 3; j++) {
                     matrix[i][j] = reader.readF16();
                 }
             }
-            opsinBias = new double[3];
+            opsinBias = new float[3];
             for (int i = 0; i < 3; i++)
                 opsinBias[i] = reader.readF16();
-            quantBias = new double[3];
+            quantBias = new float[3];
             for (int i = 0; i < 3; i++)
                 quantBias[i] = reader.readF16();
             quantBiasNumerator = reader.readF16();
@@ -74,9 +74,9 @@ public class OpsinInverseMatrix {
     }
 
     private void bakeCbrtBias() {
-        cbrtOpsinBias = new double[3];
+        cbrtOpsinBias = new float[3];
         for (int c = 0; c < 3; c++)
-            cbrtOpsinBias[c] = MathHelper.signedPow(opsinBias[c], 1D/3D);
+            cbrtOpsinBias[c] = MathHelper.signedPow(opsinBias[c], 1f/3f);
     }
 
     public OpsinInverseMatrix getMatrix(CIEPrimaries primaries, CIEXY whitePoint) {
@@ -87,17 +87,17 @@ public class OpsinInverseMatrix {
         return opsin;
     }
 
-    public void invertXYB(double[][][] buffer, double intensityTarget) {
+    public void invertXYB(float[][][] buffer, float intensityTarget) {
         if (buffer.length < 3)
             throw new IllegalArgumentException("Can only XYB on 3 channels");
         FlowHelper.parallelIterate(IntPoint.sizeOf(buffer[0]), (x, y) -> {
-            double gammaL = buffer[1][y][x] + buffer[0][y][x] - cbrtOpsinBias[0];
-            double gammaM = buffer[1][y][x] - buffer[0][y][x] - cbrtOpsinBias[1];
-            double gammaS = buffer[2][y][x] - cbrtOpsinBias[2];
-            double itScale = 255D / intensityTarget;
-            double mixL = (gammaL * gammaL * gammaL + opsinBias[0]) * itScale;
-            double mixM = (gammaM * gammaM * gammaM + opsinBias[1]) * itScale;
-            double mixS = (gammaS * gammaS * gammaS + opsinBias[2]) * itScale;
+            float gammaL = buffer[1][y][x] + buffer[0][y][x] - cbrtOpsinBias[0];
+            float gammaM = buffer[1][y][x] - buffer[0][y][x] - cbrtOpsinBias[1];
+            float gammaS = buffer[2][y][x] - cbrtOpsinBias[2];
+            float itScale = 255f / intensityTarget;
+            float mixL = (gammaL * gammaL * gammaL + opsinBias[0]) * itScale;
+            float mixM = (gammaM * gammaM * gammaM + opsinBias[1]) * itScale;
+            float mixS = (gammaS * gammaS * gammaS + opsinBias[2]) * itScale;
             for (int c = 0; c < 3; c++)
                 buffer[c][y][x] = matrix[c][0] * mixL + matrix[c][1] * mixM + matrix[c][2] * mixS;
         });
