+++ java/com/thebombzen/jxlatte/frame/features/noise/NoiseParameters.java
@@ -6,10 +6,10 @@ import com.thebombzen.jxlatte.io.Bitreader;
 
 public class NoiseParameters {
 
-    public final double[] lut = new double[8];
+    public final float[] lut = new float[8];
 
     public NoiseParameters(Bitreader reader) throws IOException {
         for (int i = 0; i < lut.length; i++)
-            lut[i] = reader.readBits(10) / 1024D;
+            lut[i] = reader.readBits(10) / 1024f;
     }
 }
